refactor: moved invite codes to class settings page
