+++ utils/validation.js
@@ -308,24 +308,31 @@ export const validateInviteCode = (req, res, studentClass, mentorClass) => {
       inviteClass = mentorClass;
       newUserRole = ClassRoles.MENTOR;
     }
-    return ClassRole.findOne({
-      userId: req.user.id,
-      classId: inviteClass.id,
-    }).then((classRole) => {
-      // If user is not enrolled in class, return the class info to add to user to it
-      if (!successfulFindOneQuery(classRole)) {
-        return new ClassRole({
-          userId: req.user.id,
-          classId: inviteClass.id,
-          role: newUserRole,
-        });
-      }
-      return sendJsonErrMessage(
-        res,
-        400,
-        "User is already enrolled in the class"
-      );
-    });
+    return (
+      ClassRole.findOne({
+        userId: req.user.id,
+        classId: inviteClass.id,
+      })
+        // Check whether or not class is closed
+        .then((classRole) =>
+          validateClassIsIncomplete(res, inviteClass.id, classRole)
+        )
+        .then((classRole) => {
+          // If user is not enrolled in class, return the class info to add to user to it
+          if (!successfulFindOneQuery(classRole)) {
+            return new ClassRole({
+              userId: req.user.id,
+              classId: inviteClass.id,
+              role: newUserRole,
+            });
+          }
+          return sendJsonErrMessage(
+            res,
+            400,
+            "User is already enrolled in the class"
+          );
+        })
+    );
   }
   return sendJsonErrMessage(res, 400, "Invalid invite code");
 };
